[2019.3.4]修复录制中时间戳的输入问题（java long类型需要转到C++的long long类型）

+++ app/src/main/cpp/egl/EGLCore.cpp
@@ -197,7 +197,7 @@ GLboolean EGLCore::buildVideoContext(ANativeWindow *window, EGLContext context) 
  * 设置显示时间戳pts
  * @param nsecs
  */
-void EGLCore::setPresentationTime(long nsecs) {
+void EGLCore::setPresentationTime(long long nsecs) {
     eglPresentationTimeANDROID(mDisplay,mSurface,nsecs);
 }
